+++ src/Microsoft.Azure.IIoT.OpcUa.Protocol/src/Extensions/StackModelsEx.cs
@@ -108,7 +108,7 @@ namespace Microsoft.Azure.IIoT.OpcUa.Protocol {
             }
             return new RolePermissionType {
                 RoleId = model.RoleId.ToNodeId(context),
-                Permissions = /*(uint)*/ model.Permissions.ToStackType()
+                Permissions = (uint)model.Permissions.ToStackType()
             };
         }
